change member id from int to long

+++ src/main/java/de/zottig/clean/persistence/dao/MemberRepository.java
@@ -9,6 +9,6 @@ import de.zottig.clean.persistence.model.Member;
 public interface MemberRepository extends JpaRepository<Member, Long> {
 
 	Member findOneByEmail(String email);
-	Member findOneById(int id);
+	Member findOneById(Long id);
 	List<Member> findByHouseholdId(Long long1);
 }
